setup JWT authentication, Identity roles, and move interfaces

+++ BLL/Settings/JwtSettings.cs
@@ -5,6 +5,6 @@ namespace BLL.Settings
         public string Key { get; set; } = string.Empty;
         public string Issuer { get; set; } = string.Empty;
         public string Audience { get; set; } = string.Empty;
-        public double DurationInMinutes { get; set; }
+        public int DurationInMinutes { get; set; }
     }
 }
